change type of id field to int. small refactoring.
