+++ ActiveCampaign/Struct/ListListOptions.cs
@@ -4,6 +4,11 @@ namespace ActiveCampaign.Struct
 {
     public class ListListOptions
     {
+        public ListListOptions()
+        {
+            Filters = new Dictionary<string, string>();
+        }
+
         public Dictionary<string, string> Filters { get; set; }
         public bool Full { get; set; }
         public bool GlobalFields { get; set; }
